+++ docs/docs/changelog.md
@@ -5,6 +5,8 @@ title: Changelog
 
 ## [Unreleased](https://github.com/lets-cli/lets/releases/tag/v0.0.X)
 
+## [0.0.56](https://github.com/lets-cli/lets/releases/tag/v0.0.56)
+
 * `[Dependency]` update go to `1.24`
 * `[Added]` support custom top-level keywords that start with `x-`
 * `[Added]` check for invalid top-level keywords during config parsing